requirements.txt
streamlit
pillow
pandas

